fix(ebpf): size of mntns/pidns filters key holders (#3337) (#3338)
Even though the namespace id is a 32-bit integer, the key size for the
mntns/pidns filters has been set to 64-bit. This caused these filters to
not work correctly as their key value was being read beyond limits.

Chore: Add integration tests for mntns/pidns filters.

commit: bb8c4c8 (main), cherry-pick

+++ pkg/ebpf/c/common/filtering.h
@@ -239,14 +239,14 @@ statfunc u64 compute_scopes(program_data_t *p)
     if (p->config->mnt_ns_filter_enabled_scopes) {
         u64 filter_out_scopes = p->config->mnt_ns_filter_out_scopes;
         u64 mask = ~p->config->mnt_ns_filter_enabled_scopes;
-        u32 mnt_id = context->mnt_id;
+        u64 mnt_id = context->mnt_id;
         res &= equality_filter_matches(filter_out_scopes, &mnt_ns_filter, &mnt_id) | mask;
     }
 
     if (p->config->pid_ns_filter_enabled_scopes) {
         u64 filter_out_scopes = p->config->pid_ns_filter_out_scopes;
         u64 mask = ~p->config->pid_ns_filter_enabled_scopes;
-        u32 pid_id = context->pid_id;
+        u64 pid_id = context->pid_id;
         res &= equality_filter_matches(filter_out_scopes, &pid_ns_filter, &pid_id) | mask;
     }
